requirements.txt
    Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
pytz==2023.3
gunicorn==21.2.0
pusher==3.3.2
python-dotenv==1.1.1